fixes #969: allow user to give DatePicker a popperContainer to wrap (#1074)
react-popper will place the popover in the same constraints as the
parent div. For cases where the parent div is somehow constrained -- a
scrollable div -- the popper will appear inside the div and will be
constrained by it to.

There are cases where you want the popover to be unconstrained it's
parent. To do this, you must place the popover in a container outside of
the constrained container.

Adding a popperContainer would allow a user to implement the following

  function CalendarContainer (props) {
    const containerId = document.getElementById('container-outside-root')

    return (
      <Portal container={containerId}>
        {props.children}
      </Portal>
    )
  }

  // somewhere else

  <DatePicker ... popperContainer={CalendarContainer} />

Which will then permit rendering the control inside the div and the
calendar outside the div (which the popper library will connect through
the magic of javascript)

demo: https://codesandbox.io/s/1ozy5ol5vl